Big changes...
A lot of work on the new router, but also additions and changes all over.  Notable changes include:

This project now relies on a custom version of Next.js.  It was just easiest to fork it and change it so it worked the way I needed it, rather than figure out how to monkey-patch it.

The `environment-loader` has been removed.  It was not playing nice with Next.js.  The `DefinePlugin` is not used in its place (look out for the `process` global variable).

Similarly, the `name-of-loader` has been altered to only retrieve names of other components.  `process.module.name` can be used to get the current module's name.

There is a currently-unused `code-patch-loader` that was originally going to be used to monkey-patch files as Webpack loaded them.  This was going to be paired with the parsing-related tools.  Unfortunately, it was not straight-forward to patch an `await` into Next.js's ES5 compiled distributable source.  The idea of using this patcher was abandoned in favor of just forking Next.js, however this loader remains here in case it become relevant in the future.