Standardizing on the term "article" for site content.
Before, the word "page" was often used to refer to the site content.  This was a little bit confusing, because the term was already used by Next.js to refer to components that were logical routes and because the original template often referred to them as "article".

Now, the term "article" refers to the site content and "page" refers to the Next.js specific concept.

All the articles have been relocated from the`pages/articles` folder to `components/articles`.  They also have their capitalized file names again.  A little machinery has been added to create nicely named `.html` files when exported.