Minor restructure of `index.jsx` & related components.
The `index` page now exposes an `appContext` prop for accessing application-wide services and resources.
The `Lightbox` component has been altered to be an `appContext` service, which gives more of its control to `index`.
The Lightbox is now rendered by the `index` page instead of the `Main` component.
A `Wrapper` component has been added to simplify `propTypes` checking.
The `imageSync` prop is now on the `appContext` object.
Various article components have been altered to reflect this change.