Additions to `components/Preloader`.
Adds the ability to get a promise for when the component begins preloading.  The promise resolves with a boolean, indicating if the preloading was successful; if the component unmounts before preloading completes, this will be `false`.
Adds an `id` property to be assigned to the rendered `div` element.
Removed usage of `tools/common.Composition`.
Other small refactors to `componentDidUpdate`.