Major redesign of `components/Nate` and related.
Now using `styled-jsx` for styling instead of SASS.  This allows use of `image-media-loader` to get the images, ensuring they exist during a build.
The refs for sprites are now stored in the `world` state.
The way sounds are accessed and played is different.  Sound playback is now handled by the `Nate` component itself rather than by `./nateLogic/nateCommon.playSound`.  A `soundPlayer` is created during the `Nate` component's initialization, and this player will be used by the game logic to playback sounds instead.
Some of the work to get a loading-spinner has been completed; the `Nate` component handles preloading of its assets, but does not yet have a mechanism to report its success or failure to its parent component.