Changes how `components/Preloadable.promised` functions.
The second argument is now an `options` object.
The `options` object can receive an `initialProps` property, the contents of which will be mixed with other properties the wrapped component receives.
The resulting component can now be provided a `render` prop which determines how the result of the promise will be rendered.