Refactor to `components/Header` and related styles.
Accepts a `logo` attribute to set the logo on the header.
Now exposes a `Header.Content` component, for providing content to the header.
Now exposes a `Header.Nav` component, for providing links on the main-menu.
These two sub-components are the only children `Header` will accept.
Change how the transitions are done in the static CSS file.
My personal modifications to the contents and style are now intended to be provided elsewhere (in the index page itself).
Later commits will reflect these changes in other locations.