Refactor to `components.Background` and related.
The `component/Background` component now has finer control over what happens when it is loading or in an error state, no longer rendering the background when it fails to load.
The `Background` component is now rendered by the `pages/_app` component.
The class-name of the background is now handled via a context, `lib/BackgroundContext`.
Components which would like to alter the background's style can do so via a `components/Background.Controller` component.
Removed unnecessary `key` props from `components/Page` and `pages/index`.  Updates sent to this context are delayed by a frame.