Refactor to `tools/common`.
`compareOwnProps` now works similarly to the way React compares props.
`disjunctiveUnion` now accepts `null` or `undefined` for each argument.  An empty object is returned in the case both are not defined.  If one is not defined but the other is, a copy of the own-properties of the defined object is returned.  Otherwise, it works the same.