Additions to `tools/extensions/common.is`.
Adds `refType` and `valueType`, for checking to see if a value is one of each.  Both `Symbol` and `String` are treated as value-types.  `null` is still recognized as a reference-type, but `undefined` is treated as a value-type.
Adds `in` for checking to see if a value is inside an array.  Inspired by CoffeeScript's `if ... in [...]` syntax.