Additions to `tools/propTypes`.
Adds `anyShape`, which is similar to the vanilla `PropTypes.shape`, except it works on any reference-type, including functions.
Adds `children` and `flatChildren` validators.  These are useful for ensuring that the children of a component have a specific element or component type as well as specific props specified.
This functionality depends on a new package, which has also been added to `package.json`.