Bugfix for `components/Preloader`.
Now correctly implements the `wait` prop.
It is no longer possible for the `onPreload` event to be skipped; even if the preloader mounts with a preload-sync in a `preloaded` state, it will now raise the `onPreload` and `onLoad` events in order when this happens.
This also fixes the possibility of no promise ever being delivered to the `promise` prop, if the `preloading` state was skipped.
Cleaned up `componentDidUpdate` and related code a little.