Configuration work.
A special override for the `transform-react-remove-prop-types` Babel plugin was added to ensure that `tools/propTypes` and related get removed.
Standardized more on using MicroMatch for path matching.
CoreJS is no longer compulsively included into the common chunk.
A couple of bugs with the `splitChunks` configuration was fixed.
Some documentation added here and there.